+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.2.5 (unreleased)
 ------------------
 
+- Fix an issue when the geolocation is empty
+  [mpeeters]
+
 - Allow to change map marker popup default content (#35).
   See ``@@geolocation-geojson-popup`` view.
   [laulaz]